% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPUtilities.R
\name{addReleaseDate}
\alias{addReleaseDate}
\title{Add release date to detections.}
\usage{
addReleaseDate(df, release.data, PTAGIS = FALSE)
}
\arguments{
\item{data.frame}{}
}
\value{
A dataframe.
}
\description{
Add release date to detections.
}
\examples{
detections.plus.release <- addReleaseDate(detections,releases)
detections.plus.release2 <- addReleaseDate(detections,releases,PTAGIS=TRUE)
}

