% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{checkReleaseDates}
\alias{checkReleaseDates}
\title{Check Release Dates}
\usage{
checkReleaseDates(df)
}
\arguments{
\item{data.frame}{}
}
\value{
data.frame
}
\description{
Check Release Dates
}
\examples{
checkReleaseDates(dataset)
}

