% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{compareReleaseDates}
\alias{compareReleaseDates}
\title{Plot listed release date vs. PTAGIS release date}
\usage{
compareReleaseDates(df, color = "release")
}
\arguments{
\item{color}{}

\item{data.frame}{}
}
\value{
data.frame
}
\description{
Plot listed release date vs. PTAGIS release date
}
\examples{
compareReleaseDates(dataset)
}

