% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPUtilities.R
\name{getMarkFileFish}
\alias{getMarkFileFish}
\title{Returns a set of tags, one from each mark file.}
\usage{
getMarkFileFish(smp)
}
\arguments{
\item{smp}{}
}
\value{
A dataframe with one column, one tag from each mark file.
}
\description{
Returns a set of tags, one from each mark file.
}
\examples{
mfile.tags <- getMarkFileFish(data)
}

