% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPUtilities.R
\name{getSampleData}
\alias{getSampleData}
\title{Get sample data}
\usage{
getSampleData(detections, smp)
}
\arguments{
\item{detections}{}

\item{smp}{}
}
\value{
data.frame
}
\description{
Get sample data
}
\examples{
srates <- getSampleData(detections,smp)
}

