% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPUtilities.R
\name{getSmallDate}
\alias{getSmallDate}
\title{Get month and day from a date}
\usage{
getSmallDate(date)
}
\arguments{
\item{date}{}
}
\value{
string
}
\description{
Get month and day from a date
}
\examples{
getSmallDate(lubridate::ymd("2012-05-01"))
}

