% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadParseData.R
\name{getSubyearlings}
\alias{getSubyearlings}
\title{Extract subyearling Fall Chinook from full detection dataset}
\usage{
getSubyearlings(df)
}
\arguments{
\item{parsed}{data.frame}
}
\value{
data.frame
}
\description{
Extract subyearling Fall Chinook from full detection dataset
}
\examples{
subs <- getSubyearlings(dataset)
}

