% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadParseData.R
\name{getYearlings}
\alias{getYearlings}
\title{Extract Yearling Fall Chinook from full detection dataset}
\usage{
getYearlings(df)
}
\arguments{
\item{data.frame}{}
}
\value{
data.frame
}
\description{
Extract Yearling Fall Chinook from full detection dataset
}
\examples{
yls <- getYearlings(dataset)
}

