% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadParseData.R
\name{parseDetections}
\alias{parseDetections}
\title{Parse CSV output from PTAGIS query}
\usage{
parseDetections(df)
}
\arguments{
\item{data.frame}{}
}
\value{
data.frame with simplified variable names and parsed dates
}
\description{
Parse CSV output from PTAGIS query
}
\examples{
parseDetections(df)
}

