% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadParseData.R
\name{parseSASDatetime}
\alias{parseSASDatetime}
\title{Convert SAS datetime string to R posix Datetime}
\usage{
parseSASDatetime(sdate)
}
\arguments{
\item{String}{}
}
\value{
Posix Datetime
}
\description{
Convert SAS datetime string to R posix Datetime
}
\examples{
parseSASDatetime("09DEC2012:01:29:11")
}

