% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{plotArrivals}
\alias{plotArrivals}
\title{Create plot of arrivals at GRJ, aggregated by day}
\usage{
plotArrivals(df, group.surrogates = TRUE, type = "area",
  scale.free = "TRUE")
}
\arguments{
\item{data.frame}{}

\item{colour.scheme}{}
}
\value{
data.frame
}
\description{
Create plot of arrivals at GRJ, aggregated by day
}
\examples{
plotArrivals(dataset)
}

