% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{plotDetections}
\alias{plotDetections}
\title{Create jitter plot of first detected date vs. last detected location}
\usage{
plotDetections(df, color = "MonitorMod")
}
\arguments{
\item{color}{}

\item{data.frame}{}
}
\value{
data.frame
}
\description{
Create jitter plot of first detected date vs. last detected location
}
\examples{
plotDetections(dataset)
}

