% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{plotMarkingTagging}
\alias{plotMarkingTagging}
\title{Plot bar chart to check for potential matches for untagged releases}
\usage{
plotMarkingTagging(releases, mode = "marking")
}
\arguments{
\item{releases}{}

\item{mode}{}
}
\description{
Plot bar chart to check for potential matches for untagged releases
}
\examples{
plotMarkingTagging(releases)
}

