% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{plotOperations}
\alias{plotOperations}
\title{Plot Operational Dates}
\usage{
plotOperations(ops, size = 5)
}
\arguments{
\item{ops}{}

\item{size}{}
}
\description{
Plot Operational Dates
}
\examples{
ops <- read.csv("operations.csv")
plotOperations(ops)
}

