% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPPlottingFunctions.R
\name{plotSampleRates}
\alias{plotSampleRates}
\title{Plot sample rates}
\usage{
plotSampleRates(sample.data, facet = FALSE)
}
\arguments{
\item{data.frame}{}

\item{data.frame}{}
}
\value{
data.frame
}
\description{
Plot sample rates
}
\examples{
plotSampleRates(dataset)
}

