% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PTAGIS.MRRSites}
\alias{PTAGIS.MRRSites}
\title{PTAGIS MRR Sites}
\format{A data frame with 336 rows and 18 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.ptagis.org/}
}
\usage{
PTAGIS.MRRSites
}
\description{
A dataframe with the PIT-tag MRR Sites
}
\keyword{datasets}

