% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SMPUtilities.R
\name{surrogates}
\alias{surrogates}
\title{Get surrogates (i.e. GATE1 etc.).}
\usage{
surrogates(df, keep = TRUE)
}
\arguments{
\item{data.frame}{}
}
\value{
A dataframe.
}
\description{
Get surrogates (i.e. GATE1 etc.).
}
\examples{
surrogates <- surrogates(data)
}

